Added layers PlayerObjects and Target Objects
These layers can be accessed by clicking on an object in the hierarchy, going into Sprite Renderer, clicking on additional settings and there you will find the layers next Sorting Layers. The background is on default, the ball is on "PlayerObjects" and the trash can is on "Target Objects". "Target Objects" is in front of all other layers. Default is under all other layers. "Player Objects" is in between the two.